third_party/WebKit/LayoutTests/animations/animationworklet/cancel-non-accelerated-property-expected.txt

CONSOLE MESSAGE: line 49: background-color for the first target is: rgb(0, 255, 0)
